+++ src/help/mod.rs
@@ -4,4 +4,7 @@ mod help_search_modes;
 mod help_state;
 mod help_verbs;
 
-pub use help_state::HelpState;
+pub use {
+    help_state::HelpState,
+    help_search_modes::*,
+};